# .github/workflows/openAPI.yml
openapi: '3.0.2'
info:
  title: My Wedding Day
  description: Alta Project 
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://app.swaggerhub.com/apis/Nathannov24/WeddingDay/1.0
...
paths:
...
  /login/organizer:
      post:
        tags:
          - Auth
        summary: Get access to account organizer
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/requestLogin'
              examples:
                userData1:
                  description: "Example data user"
                  value: {
                    email: "alta@alterra.id",
                    password: "inipassword"
                  }
        responses:
          '200':
            description: Success Response
            content:
              application/json:
                examples:
                  responseData1:
                    description: success
                    value: {
                      id: 1,
                      message: "Success Login",
                      name: "a",
                      role: "organizer",
                      status: "success",
                      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MzgzMjYxODAsInVzZXJJZCI6MSwidXNlcm5hbWUiOiJhZG1pbiJ9.AebFR-oQjUSOMez2ucDWkiMrS2eQIPmcYm5c71qZ_co"
                    }
          '401':
            description: Failed Response
            content:
              application/json:
                examples:
                  responseData1:
                    description: Email & Password Doesn't Match
                    value: {
                      code: 400,
                      status: "failed",
                      message: "Login Failed. Email and Password incorrect."
                    }  
  ####################################
  #   HOME
  #################################### 
...
  /register/organizer:
    post:
      tags:
        - Organizer
      summary: Create Account Organizer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestRegisterOrganizer'
            examples:
              userData1:
                description: "Example data organizer"
                value: {
                  WoName: "a",
                  email: "alta@alterra.id",
                  PhoneNumber: "12345678",
                  password: "a",
                  city: "depok",
                  address: "Jl. Raya Bedahan, Bedahan, Kec. Sawangan, Kota Depok, Jawa Barat 16519"
                }
      responses:
        '201':
          description: Success Response
          content:
            application/json:
              examples:
                responseData1:
                  description: success
                  value: {
                    message: "Success Create New Organizer",
                    status: "success"
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: Email Has Been Used
                  value: {
                    message: "Email was used, try another email",
                    status: "failed"
                  }
...
components:
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  ####################################
  #   SCHEMAS
  ####################################
  schemas:
    requestLogin:
      type: object
      properties:
        "email":
          type: string
          minLength: 1
        "password":
          type: string
          minLength: 1
...
    requestRegisterOrganizer:
      type: object
      properties: {
        "WoName":{
          type: string
        },
        "email":{
          type: string
        },
        "password":{
          type: string
        },
        "PhoneNumber":{
          type: string
        },
        "city":{
          type: string
        },
        "address":{
          type: string
        },
        "urlweb":{
          type: string
        },
        "description":{
          type: string
        },
        "logo":{
          type: string
        }
      }
...